Separation of AI from sheep
We separated the AI part of the game from sheep so enemy can use the module now.
