+++ StealMySheep/AI/src/main/nbm/manifest.mf
@@ -1,0 +1,3 @@
+Manifest-Version: 1.0
+OpenIDE-Module-Localizing-Bundle: stealmysheep/ai/Bundle.properties
+
