+++ StealMySheep/AI/src/main/resources/stealmysheep/ai/Bundle.properties
@@ -1,0 +1,6 @@
+#Localized module labels. Defaults taken from POM (<name>, <description>, <groupId>) if unset.
+#OpenIDE-Module-Name=
+#OpenIDE-Module-Short-Description=
+#OpenIDE-Module-Long-Description=
+#OpenIDE-Module-Display-Category=
+#Sun Apr 19 12:07:13 CEST 2020
