+++ StealMySheep/application/pom.xml
@@ -69,6 +69,11 @@
             <artifactId>Weapon</artifactId>
             <version>${project.version}</version>
         </dependency>
+        <dependency>
+            <groupId>${project.groupId}</groupId>
+            <artifactId>AI</artifactId>
+            <version>${project.version}</version>
+        </dependency>
     </dependencies>
